1. move Zend/Stdlib/ArrayObject/PhpReferenceCompatibility.php into Zend/Stdlib/ArrayObject.php and remove the Zend/Stdlib/ArrayObject/PhpReferenceCompatibility.php 2. move Zend/Session/Storage/SessionArrayStorage/PhpReferenceCompatibility.php into Zend/Session/Storage/SessionArrayStorage.php and remove Zend/Session/Storage/SessionArrayStorage/PhpReferenceCompatibility.php
