--- library/Zend/Session/Storage/SessionArrayStorage/PhpReferenceCompatibility.php
@@ -1,40 +1,0 @@
-<?php
-/**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
- */
-
-namespace Zend\Session\Storage\SessionArrayStorage;
-
-use Zend\Session\Storage\AbstractSessionArrayStorage;
-
-/**
- * PHP 5.3.4 and greater variant of SessionArrayStorage
- */
-abstract class PhpReferenceCompatibility extends AbstractSessionArrayStorage
-{
-    /**
-     * Get Offset
-     *
-     * @param  mixed $key
-     * @return mixed
-     */
-    public function &__get($key)
-    {
-        return $_SESSION[$key];
-    }
-
-    /**
-     * Offset Get
-     *
-     * @param  mixed $key
-     * @return mixed
-     */
-    public function &offsetGet($key)
-    {
-        return $_SESSION[$key];
-    }
-}
